ci: moved some tests to a cassndra only file
